+++ src/main/java/com/example/oatnote/user/service/UserRepository.java
@@ -1,9 +1,9 @@
-package com.example.oatnote.user;
+package com.example.oatnote.user.service;
 
 import org.springframework.data.mongodb.repository.MongoRepository;
 import java.util.Optional;
 
-import com.example.oatnote.user.models.User;
+import com.example.oatnote.user.service.models.User;
 
 public interface UserRepository extends MongoRepository<User, String> {
